chore: Improve build runtime - best practice
This is an example of a build best practice. By copying the package.json
file before copying in any other files, the dependencies in the
package.json file are ignored if cached and there have been no chnages.
This way the build runs from cached sources instead of rebuilding
everything.